requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
black==23.9.1
boto3==1.26.151
botocore==1.29.151
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.2.0
ci-info==0.3.0
click==8.1.7
coloredlogs==15.0.1
configobj==5.0.8
configparser==6.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
dataclasses-json==0.5.14
defusedxml==0.7.1
deprecation==2.1.0
Django==4.2.7
django-cors-headers==3.14.0
django-storages==1.13.2
django-templated-mail==1.1.1
django-webpack-loader==1.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
dnspython==2.4.2
etelemetry==0.3.1
filelock==3.13.0
frozenlist==1.4.0
future==0.18.3
fuzzywuzzy==0.18.0
ghostscript==0.7
greenlet==2.0.2
gunicorn==21.2.0
httplib2==0.22.0
humanfriendly==10.0
idna==3.4
img2pdf==0.4.4
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.348
langchain-core==0.0.12
langsmith==0.0.69
Levenshtein==0.23.0
loguru==0.7.2
looseversion==1.3.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
munch==4.0.0
mypy-extensions==1.0.0
networkx==3.2.1
nibabel==5.1.0
nipype==1.8.6
numexpr==2.8.6
numpy==1.23.2
oauthlib==3.2.2
ocrmypdf==14.0.0
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.2
pandas==2.1.2
pathlib==1.0.1
pathspec==0.11.2
pdfminer.six==20201018
pikepdf==8.6.0
Pillow==10.1.0
pinecone-client==2.2.1
platformdirs==3.10.0
pluggy==0.13.1
prov==2.0.0
psycopg2-binary==2.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydot==1.4.2
Pygments==2.16.1
PyJWT==2.8.0
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pyOpenSSL==23.3.0
pyparsing==3.1.1
pypdf==3.16.1
pypdfium2==4.23.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.23.0
python3-openid==3.2.0
pytwitter==0.5
pytz==2023.3.post1
pyxnat==1.6
PyYAML==6.0.1
rapidfuzz==3.5.2
rdflib==7.0.0
regex==2023.8.8
reportlab==4.0.6
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
s3transfer==0.6.2
scipy==1.11.3
service-identity==23.1.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
tenacity==8.2.3
tesseract==0.1.3
tiktoken==0.5.1
tqdm==4.66.1
traits==6.3.2
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
yarl==1.9.2
